requirements.txt
opencv-python==4.11.0.86
numpy==2.1.1
Pillow==10.4.0 